// frontend/libs/components/video-editor/src/lib/timeline/types.ts
interface BaseTrack {
  id: string;
  name: string;
}

export interface VideoTrack extends BaseTrack {
  type: "video";
  elements: (VideoElement | ImageElement)[];
  muted: boolean;
  hidden: boolean;
}

export interface TextTrack extends BaseTrack {
  type: "text";
  elements: TextElement[];
  hidden: boolean;
}

export interface AudioTrack extends BaseTrack {
  type: "audio";
  elements: AudioElement[];
  muted: boolean;
}

export interface GraphicTrack extends BaseTrack {
  type: "graphic";
  elements: (StickerElement | GraphicElement)[];
  hidden: boolean;
}

export interface EffectTrack extends BaseTrack {
  type: "effect";
  elements: EffectElement[];
  hidden: boolean;
}

export type TimelineTrack =
  | VideoTrack
  | TextTrack
  | AudioTrack
  | GraphicTrack
  | EffectTrack;

export type OverlayTrack = VideoTrack | TextTrack | GraphicTrack | EffectTrack;

export interface SceneTracks {
  overlay: OverlayTrack[];
  main: VideoTrack;
  audio: AudioTrack[];
}
// ...
export type AudioElement = UploadAudioElement | LibraryAudioElement;

interface BaseTimelineElement {
  id: string;
  name: string;
  duration: MediaTime;
  startTime: MediaTime;
  trimStart: MediaTime;
  trimEnd: MediaTime;
  sourceDuration?: MediaTime;
  animations?: ElementAnimations;
  params: ParamValues;
}

export interface VideoElement extends BaseTimelineElement {
  type: "video";
  mediaId: string;
  isSourceAudioEnabled?: boolean;
  hidden?: boolean;
  retime?: RetimeConfig;
  effects?: Effect[];
  masks?: Mask[];
}

export interface ImageElement extends BaseTimelineElement {
  type: "image";
  mediaId: string;
  hidden?: boolean;
  effects?: Effect[];
  masks?: Mask[];
}